Started to seperate adventure runtime and builder projects. Fixed a timer memleak. Fixed some memleaks in Engine.
git-svn-id: svn://svn.code.sf.net/p/captainsengine/code/trunk@336 dc34a4ac-3e42-4537-918b-2dea8b01751e

+++ Engine/gui/listbox.cpp
@@ -85,7 +85,7 @@ bool ListBox::isClicked(const ::CGE::Vector2D& pos){
       && pos.y < pos_.y+span_.y/2){
     //is scrolling neccessary
     int maxy = span_.y / 20;
-    if (scroll_+maxy >= entries_.size())
+    if (scroll_+maxy >= (int)entries_.size())
       return false;
     scroll_++;
     return false;
